github_utility/github_cli/requirements.txt
PyGithub==2.5.0
python_dateutil==2.9.0.post0
typer==0.15.0
